+++ pkg/filter/http/remote/resolver/dubboresolver/dubbo_resolver_test.go
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package dubbo
+package dubboresolver
 
 import (
 	"context"
